refactor: separate Tiling and TilingProvider
